Include application for each service

+++ services/bank/src/main/java/org/alma/services/bank/api/services/IDomainService.java
@@ -4,7 +4,7 @@ package org.alma.services.bank.api.services;
  * Created by jeremy on 23/11/15.
  */
 public interface IDomainService {
-    public int CreditAccount(int amount);
-    public int GetCredit();
-    public boolean MakeOnlinePayment(int amount);
+    public double CreditAccount(double amount);
+    public double GetCredit();
+    public boolean MakeOnlinePayment(double amount);
 }
